# api/swagger.yml
...
consumes:
- application/json
produces:
- application/json
swagger: "2.0"
info:
  description: API documentation
  title: github.com/kashguard/go-mpc-vault
  version: 0.1.0
paths:
...
  /api/v1/organizations/{orgId}/members/{userId}:
    delete:
      description: Remove member from organization
      tags:
      - organization
      summary: Remove Organization Member
      operationId: DeleteOrganizationMemberRoute
      parameters:
      - type: string
        name: orgId
        in: path
        required: true
      - type: string
        name: userId
        in: path
        required: true
      responses:
        "200":
          description: Member removed
...
  /api/v1/requests/{requestId}/approve:
    post:
      security:
      - Bearer: []
      tags:
      - signing
      summary: Approve or reject a signing request
      operationId: PostApproveSigningRequest
      parameters:
      - type: string
        name: requestId
        in: path
        required: true
      - name: Payload
        in: body
        schema:
          $ref: '#/definitions/approveSigningRequestPayload'
      responses:
        "200":
          description: Request Updated
          schema:
            $ref: '#/definitions/approveSigningResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Request Not Found
...
definitions:
...
  approveSigningRequestPayload:
    type: object
    required:
    - action
    properties:
      action:
        type: string
        enum:
        - approve
        - reject
      authenticator_data:
        description: Base64 encoded WebAuthn Authenticator Data
        type: string
        format: byte
      client_data_json:
        description: Base64 encoded WebAuthn Client Data JSON
        type: string
        format: byte
      comment:
        type: string
      credential_id:
        description: Base64 encoded WebAuthn Credential ID
        type: string
        format: byte
      signature:
        description: Base64 encoded WebAuthn Assertion Signature
        type: string
        format: byte
  approveSigningResponse:
    type: object
    properties:
      current_approvals:
        type: integer
      required_approvals:
        type: integer
      status:
        type: string
...
securityDefinitions:
  Bearer:
    description: |-
      Access token for application access, **must** include "Bearer " prefix.
      Example: `Bearer b4a94a42-3ea2-4af3-9699-8bcbfee6e6d2`
    type: apiKey
    name: Authorization
    in: header
    x-keyPrefix: 'Bearer '
  Management:
    description: Management secret, used for monitoring and infrastructure related
      calls
    type: apiKey
    name: mgmt-secret
    in: query